refactor: 重构后端自动注入方式并清理无用代码
