+++ backend_todolist/src/main/java/com/todolistbackend/controller/PhotoController.java
@@ -1,8 +1,8 @@
 package com.todolistbackend.controller;
 
 import com.todolistbackend.service.PhotoService;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,9 +13,9 @@ import java.io.IOException;
 
 @RestController
 @Slf4j
+@RequiredArgsConstructor
 public class PhotoController {
-    @Autowired
-    private PhotoService photoService;
+    private final PhotoService photoService;
 
     @PostMapping("/setphoto")
     public void setPhoto(MultipartFile file, Integer userid) throws IOException {
