+++ backend_todolist/src/main/java/com/todolistbackend/interceptor/JWTInterceptor.java
@@ -1,6 +1,7 @@
 package com.todolistbackend.interceptor;
 
 import com.todolistbackend.utils.TokenUtils;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
@@ -8,7 +9,7 @@ import javax.servlet.http.HttpServletResponse;
 
 public class JWTInterceptor implements HandlerInterceptor {
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+    public boolean preHandle(HttpServletRequest request, @NotNull HttpServletResponse response, @NotNull Object handler) {
         String token = request.getHeader("token");
         return TokenUtils.checkToken(token);
     }
