+++ backend_todolist/src/main/java/com/todolistbackend/service/MailService.java
@@ -1,0 +1,5 @@
+package com.todolistbackend.service;
+
+public interface MailService {
+    void sendSimpleMail(String to, String subject, String content);
+}
