--- backend_todolist/src/main/java/com/todolistbackend/service/impl/MailService.java
@@ -1,62 +1,0 @@
-package com.todolistbackend.service.impl;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.FileSystemResource;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.stereotype.Service;
-
-import javax.mail.MessagingException;
-import javax.mail.internet.MimeMessage;
-import java.io.File;
-
-@Service("mailService")
-public class MailService {
-    @Value("${spring.mail.username}")
-    private String from;
-
-    @Autowired
-    private JavaMailSender mailSender;
-
-    public void sendSimpleMail(String to, String subject, String content) {
-        SimpleMailMessage message = new SimpleMailMessage();
-        message.setFrom(from);
-        message.setSubject(subject);
-        message.setTo(to);
-        message.setText(content);
-        try {
-            mailSender.send(message);
-        } catch (Exception e) {
-            System.out.println("发送定时邮件失败");
-            e.printStackTrace();
-        }
-
-    }
-
-    public void sendHtmlMail(String to, String subject, String content) throws MessagingException {
-        MimeMessage message = mailSender.createMimeMessage();
-        MimeMessageHelper helper = new MimeMessageHelper(message, true);
-        helper.setFrom(from);
-        helper.setTo(to);
-        helper.setSubject(subject);
-        helper.setText(content);
-        mailSender.send(message);
-    }
-
-    public void sendAttachmentsMail(String to, String subject, String content, String filePath) throws MessagingException {
-        MimeMessage message = mailSender.createMimeMessage();
-        MimeMessageHelper helper = new MimeMessageHelper(message, true);
-        helper.setFrom(from);
-        helper.setTo(to);
-        helper.setSubject(subject);
-        helper.setText(content);
-
-        FileSystemResource file = new FileSystemResource(new File(filePath));
-        String fileName = file.getFilename();
-        helper.addAttachment(fileName, file);
-
-        mailSender.send(message);
-    }
-}
