+++ backend_todolist/src/main/java/com/todolistbackend/service/impl/MailServiceImpl.java
@@ -1,0 +1,32 @@
+package com.todolistbackend.service.impl;
+
+import com.todolistbackend.service.MailService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Service;
+
+@Service("mailService")
+@RequiredArgsConstructor
+@Slf4j
+public class MailServiceImpl implements MailService {
+    private final JavaMailSender mailSender;
+
+    @Value("${spring.mail.username}")
+    private String from;
+
+    public void sendSimpleMail(String to, String subject, String content) {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setFrom(from);
+        message.setSubject(subject);
+        message.setTo(to);
+        message.setText(content);
+        try {
+            mailSender.send(message);
+        } catch (Exception e) {
+            log.warn("发送定时邮件失败");
+        }
+    }
+}
